pyproject.toml
[project]
name = "medical-records-downloader"
version = "0.1.0"
description = "Download medical records from CUF and Joaquim Chaves Saúde portals"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
cuf-mcp = "mcp_server:main"

[tool.uv]
dev-dependencies = []
